refactor(ai/openai): move OpenAI provider to subpackage and update factory usage
