+++ ai/factory.go
@@ -3,6 +3,7 @@ package ai
 import (
 	"fmt"
 
+	"github.com/wert2all/ai-commit/ai/openai"
 	"github.com/wert2all/ai-commit/project"
 )
 
@@ -27,7 +28,7 @@ func NewProvider(config Config) (Provider, error) {
 	case ProviderMistral:
 		return NewMistralProvider(config.APIKey, config.Model), nil
 	case ProviderOpenAI:
-		return NewOpenAIProvider(config.APIKey, config.Model), nil
+		return openai.NewGPTProvider(config.APIKey, config.Model), nil
 	case ProviderClaude:
 		return NewClaudeProvider(config.APIKey, config.Model), nil
 	case ProviderGemini:
